posyandu/posbindu/risk-assessment-service/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
django-filter==23.3
numpy==1.24.3
